Update README.md
Keeping in consideration the progression analysis done on the data, this web app is created to ensure a proper interface that can be easily utilized by individuals who want to check if they suffer from dementia or not. It is an easy and simple interface that allows the subjects to enter readings according to a scale and informed if they have dementia or not. The web app also has functionality for users to upload an image. The uploaded image is processed and provides a result by identifying the facial emotion shown in the image.